feat(asr): hallucination filter diagnostics record real per-run counts
Previously _apply_hallucination_stats wrote hardcoded placeholders
(version "v1", dropped=1, {"training_phrase": 1}) regardless of actual
filtering. Now the filter returns per-reason drop counts, accumulates
them across VAD chunks, resets per run to avoid cross-task leakage when
the Pipeline instance is reused by WebUI tasks, and PipelineStats
declares the three diagnostic fields as proper dataclass fields with
round-trip serialization (including the new hallucination_drop_reasons).

Also: type annotations for cli.run overrides (dict[str, Any]) and
golden_gate._rate (float denominators).

Verified: 1633 passed, 3 skipped; ruff check/format clean on all
changed files. Design note: docs/幻觉过滤诊断真实计数落地-2026-09-17.md

+++ vocal_subtitle/quality/golden_gate.py
@@ -39,7 +39,7 @@ def _matched(
     return classify_expected_match(expected, predicted)["matched"]
 
 
-def _rate(count: int, denominator: int) -> float:
+def _rate(count: float, denominator: float) -> float:
     return round(count / denominator, 6) if denominator else 0.0
